feat(editable'): add base components and stuff
