+++ src/layouts/components/Editable/index.ts
@@ -1,0 +1,1 @@
+export * from "./Editable"
