--- src/layouts/components/EditableSidebar/index.ts
@@ -1,1 +1,0 @@
-export * from "./EditableSidebar"
